+++ slim/support/peewee/view.py
@@ -1,4 +1,6 @@
 import logging
+import psycopg2
+
 import peewee
 from typing import Type, Tuple, List, Iterable, Union
 
@@ -21,7 +23,7 @@ from ...base.sqlquery import SQL_TYPE, SQLForeignKey, SQL_OP, SQLQueryInfo, SQLQ
     SQLValuesToWrite, UpdateInfo
 from ...exception import RecordNotFound, AlreadyExists, ResourceException, NotNullConstraintFailed
 from ...base.permission import DataRecord, Permissions
-from ...utils import to_bin, pagination_calc, dict_filter
+from ...utils import to_bin, pagination_calc, dict_filter, get_bytes_from_blob
 from ...base.view import AbstractSQLView, AbstractSQLFunctions, ViewOptions
 
 logger = logging.getLogger(__name__)
@@ -61,6 +63,8 @@ class PeeweeDataRecord(DataRecord):
         for name, v in model_to_dict(self.val, recurse=False).items():
             if isinstance(fields[name], peewee.ForeignKeyField):
                 name = name + '_id'
+            elif isinstance(fields[name], peewee.BlobField):
+                v = get_bytes_from_blob(v)
             if self.selected != ALL_COLUMNS and (self.selected and (name not in self.selected)):
                 continue
             data[name] = v
@@ -98,7 +102,7 @@ class PeeweeContext:
         db = self.db
         if isinstance(exc_val, peewee.IntegrityError):
             db.rollback()
-            if exc_val.args[0].startswith('duplicate key'):
+            if exc_val.args[0].startswith('duplicate key') or '唯一约束' in exc_val.args[0]:
                 raise AlreadyExists()
             elif exc_val.args[0].startswith('NOT NULL constraint failed'):
                 raise NotNullConstraintFailed()
@@ -161,17 +165,16 @@ class PeeweeSQLFunctions(AbstractSQLFunctions):
     async def select_page(self, info: SQLQueryInfo, size=1, page=1) -> Tuple[Tuple[DataRecord, ...], int]:
         q = self._make_select(info)
         count = q.count()
-        if count == 0: raise RecordNotFound(self.vcls.table_name)
+
+        # 0.4.2: list api does not return NOT_FOUND anymore
+        # if count == 0: raise RecordNotFound(self.vcls.table_name)
 
         if size == -1:
             page = 1
             size = count
 
-        try:
-            func = lambda item: PeeweeDataRecord(None, item, view=self.vcls)
-            return tuple(map(func, q.paginate(page, size))), count
-        except self._model.DoesNotExist:
-            raise RecordNotFound(self.vcls.table_name)
+        func = lambda item: PeeweeDataRecord(None, item, view=self.vcls)
+        return tuple(map(func, q.paginate(page, size))), count
 
     def _build_write_condition(self, records: Iterable[DataRecord]):
         records_pk = []