+++ slim/utils/__init__.py
@@ -4,6 +4,8 @@ import re
 import string
 import sys
 import time
+from typing import Optional
+
 from .async_run import *
 from .binhex import to_bin, to_hex
 from .cls_init import MetaClassForInit
@@ -41,12 +43,14 @@ def dict_filter_inplace(obj, keys):
         del obj[i]
 
 
-def get_bytes_from_blob(val) -> bytes:
+def get_bytes_from_blob(val) -> Optional[bytes]:
     """ 不同数据库从blob拿出的数据有所差别，有的是memoryview有的是bytes """
     if isinstance(val, bytes):
         return val
     elif isinstance(val, memoryview):
         return val.tobytes()
+    elif val is None:
+        return None
     else:
         raise TypeError('invalid type for get bytes')